Move manage apprentices action to apprentices ctr
